dnsdist: Use an `int` since it's what `recvmmsg` returns
(cherry picked from commit 0014f8e2d23d2d812c4db52d3e15fcc370f67f5a)

+++ pdns/dnsdistdist/dnsdist.cc
@@ -2136,7 +2136,7 @@ static void MultipleMessagesUDPClientThread(ClientState* clientState)
     /* reset the IO vector, since it's also used to send the vector of responses
        to avoid having to copy the data around
        No need to reset the parts that have not been used, though. */
-    for (size_t idx = 0; idx < static_cast<size_t>(msgsGot); idx++) {
+    for (int idx = 0; idx < msgsGot; idx++) {
       auto& slot = recvData[idx];
       /* only resize if the buffer is actually smaller than expected */
       if (slot.packet.size() < initialBufferSize) {
